requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
typer>=0.9.0
pydantic>=2.5.0
python-dateutil>=2.8.2
openai>=1.40.0
python-dotenv>=1.0.0
